chore: Merges useSetLocation and useRegionModal into useRegionManager
