+++ packages/core/src/Layout.tsx
@@ -1,12 +1,12 @@
 import { useMemo, type PropsWithChildren, type ReactElement } from 'react'
 
-import { useRegionModal } from './components/region/RegionModal/useRegionModal'
+import { useRegionManager } from './components/region/RegionModal/useRegionManager'
 import { usePageViewEvent } from './sdk/analytics/hooks/usePageViewEvent'
 
 function Layout({ children }: PropsWithChildren) {
   const props = useMemo(() => (children as ReactElement)?.props, [])
   usePageViewEvent(props)
-  useRegionModal()
+  useRegionManager()
 
   return <>{children}</>
 }
