+++ packages/core/src/components/region/RegionButton/RegionButton.tsx
@@ -5,12 +5,12 @@ import { deliveryPromise, session as initialSession } from 'discovery.config'
 import { useSession } from 'src/sdk/session'
 import { textToTitleCase } from 'src/utils/utilities'
 
-import { useRegionModal } from '../RegionModal/useRegionModal'
+import { useRegionManager } from '../RegionModal/useRegionManager'
 
 function RegionButton({ icon, label }: { icon: string; label: string }) {
   const { openModal, openPopover } = useUI()
   const { city, postalCode } = useSession()
-  const { isValidationComplete } = useRegionModal()
+  const { isValidationComplete } = useRegionManager()
   const regionButtonRef = useRef<HTMLButtonElement>(null)
 
   const defaultPostalCode =
