+++ packages/core/src/components/region/RegionModal/useRegionManager.ts
@@ -1,14 +1,17 @@
-import { useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 
 import type { Session } from '@faststore/sdk'
-import { sessionStore, validateSession } from 'src/sdk/session'
+import { sessionStore, useSession, validateSession } from 'src/sdk/session'
 
-interface UseSetLocationParams {
+import { useUI } from '@faststore/ui'
+import { deliveryPromise } from 'discovery.config'
+
+interface UseRegionManagerParams {
   input: string
   setInput: (value: string) => void
   resetInputField: () => void
   setLocation: (
-    postalCode: string | undefined,
+    postalCode: string | null,
     inputFieldErrorMessage: string,
     session: Session,
     onSuccess?: () => void
@@ -16,12 +19,29 @@ interface UseSetLocationParams {
   errorMessage: string
   setErrorMessage: (value: string) => void
   loading: boolean
+  isValidationComplete: boolean
+  openRegionModal: () => void
+  postalCode: string | null
 }
 
-export function useSetLocation(): UseSetLocationParams {
+export function useRegionManager(): UseRegionManagerParams {
   const [input, setInput] = useState<string>('')
   const [errorMessage, setErrorMessage] = useState<string>('')
   const [loading, setLoading] = useState<boolean>(false)
+
+  // State to track if validation is complete
+  const [isValidationComplete, setValidationComplete] = useState(false)
+  const { isValidating } = useSession()
+  const prevIsValidating = useRef(isValidating)
+
+  const { openModal: displayRegionModal } = useUI()
+  const { postalCode } = sessionStore.read()
+
+  const openRegionModal = () => {
+    if (!postalCode) {
+      displayRegionModal()
+    }
+  }
 
   const resetInputField = () => {
     setInput('')
@@ -59,6 +79,26 @@ export function useSetLocation(): UseSetLocationParams {
     }
   }
 
+  // Effect to handle when isValidating changes from true to false
+  useEffect(() => {
+    if (!deliveryPromise.enabled) {
+      return
+    }
+
+    // Check if validation has completed (isValidating changed from true to false)
+    if (prevIsValidating.current && !isValidating) {
+      setValidationComplete(true)
+
+      // If the postal code is not set and is mandatory, open the region modal
+      if (deliveryPromise.mandatory) {
+        openRegionModal()
+      }
+    }
+
+    // Update the previous value of isValidating
+    prevIsValidating.current = isValidating
+  }, [openRegionModal])
+
   return {
     input,
     setInput,
@@ -67,5 +107,8 @@ export function useSetLocation(): UseSetLocationParams {
     errorMessage,
     setErrorMessage,
     loading,
+    isValidationComplete,
+    openRegionModal,
+    postalCode,
   }
 }
